Implement method for loading the existing pki in ipfs, and storing it in the peergos user.
To do this I added a new optional field on WriterData which is a map of
named owned keys. This allows us to rotate the PKI key without any MITM
possibility because the clients will have the 'peergos' user identity
key built in. The pki key is independent of the 'peergos' user keys.