Push the space checking key filter out of the UserService so that it only wraps the local dht, not the p2p version.
In principle this should allow users to write through one node and for
the write to be proxied to their storage node. However it seems that
ipfs closes the p2p stream before we can send the response back.

Also fix a bug when changing keys and using the wrong (new) signer as
the owner.