Fixes for access via ipfs (p2p stream or static site)
IPFS doesn't expose id() over the gateway, so in this case return a
dummy node id (all calls will be proxied).

In the web ui make all ajax calls either absolute or relative. In the
case that we are talking to a peergos server (either directly, or over
a p2p stream) then all calls are relative, which also means the peergos
server can be hosted on a non root path. If we are using the web-ui
directly from an ipfs gateway statically, then all urls need to be
absolute.