[cli] auto-complete update
Simplified the auto-complete implemetation.

Also updated its behaviour to extend auto-complete options beyond the
children of pwd st/ now if you do, for example,

(ls|get|put|rm) /user/shared/<TAB>

you will see the auto-complete options for
/user/shared/friend_user1
/user/shared/friend_user2

etc. and so on for other  sub (or relative) directories.

Also fixed ls to show individual files (if the arg. is a file) rather
than just the directory listing if it's a directory).