Fix a few issues found by the fuzzer.
If we had only read access to a dir but write access to a descendant,
we wre not getting the correct (writable and with link) descendant in
several cases.

The fix involved keeping a reference to the trienode in the FileWrapper.

Cleaning a file was also removing any writer present in the parent link.