Total rewrite of sharedwith cache
Now the sharing state of each file is stored in a mirror directory tree.
This means we no longer need to build a sharedWith cache during login.
Previously building this cache was O(things we've shared) and took ram with same O().

Now the contribution to login is 0. When we browse to a directory we in a UI we
need to retrieve a single file to show the sharing state for all child items.