Implement group sharing.
This gives an O(1) way to share something with an arbitrarily large group of people.
You can define your own groups.

Members of a group cannot see:
 1) other members in the group
 2) how many members there are
 3) the name of the group

 The most important groups are two predefined ones for friends and followers

Implement removal of users from groups.
This is expensive because we need to revoke access to everyting ever shared with the group.

Also use this revocation during removing a friend/follower.

Add a group sharing benchmark

Add convenience methods for sharing wth friends or followers