This is a huge rewrite of Chat implementation
The aim was to enable atomic merging of new messages.

Most chat related classes have been made immutable.
Invite messages now explicitly include the recipient Id.
Chat's now include the chatUid in their state explicitly.
Applying messages to a Chat results in a ChatUpdate which can be combined with others, or applied atomically.

This also implied a large rework of the SharedWithCache, to ensure no nested transactions.

Still need to restore and update 2 of the Chat unit tests.