Implement 3rd party identity proofs
These are just a signed statement saying that you control
this peergos account and this 3rd party account which is posted
to the other service and also stored in Peergos.

We take it a step further and allow the proof to be encrypted.
This allows you to post a proof to the 3rd party service (which reveals nothing)
and then selectively reveal it by sharing the key. This can be simply sharing
the proof file in Peergos with friends, or out of band.