First attempt at separating identity from login.
This moves the UserStaticData from a public field to a secret blob which you need to auth with your server to retrieve.
The auth is by the signing key pair generated at login from your password + username + salt.

To set/update the login data sign it with the identity key pair.

ToDo: retrieve this login data during migration/mirroring.