Implement diff sync
This adds a new custom tree hash to file metadata. (And
drops blake3 due to poor JS libs and webcrypto absence)

Each plaintext chunk is hashed with sha256, this is then
made into a 1024-ary tree by hashing until the root.
Wrapping each layer in cbor. This means a 5 PiB file will
have an overhead of 96 KiB int eh first chunk's metadata.
So less than the max thumbnail size.

This will make is easy to do bulk uploads in browser with
hashes present, and thus not conflict with a concurrent sync
client that is setting the hashes.